Add files via upload
Comparision of Two Linked Lists
You’re given the pointer to the head nodes of two linked lists. Compare the data in the nodes of the linked lists to check if they are equal. If all data attributes are equal and the lists are the same length, return 1 . Otherwise, return .0
example:
list1=1->2->3->null
list2=1->2->3->4->null
The two lists have equal data attributes for the first  nodes.  is longer, though, so the lists are not equal. Return 0